IPython Console: Move its shell widget to widgets/jupyter_qtconsole
